requirements.txt
# requirements.txt
tensorflow==2.13.0
keras==2.13.1
torch
torchvision
opencv-python==4.8.0.74
Pillow==9.5.0
scikit-image==0.20.0
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
gradio==3.35.2
tqdm==4.65.0
albumentations==1.3.0
